agregar venta, promocion y cache

+++ src/main/java/com/utp/casa_europa/dtos/DetalleVentaDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DetalleVentaDto {
-    private Integer idProducto;
+    private Long idProducto;
     private Integer cantidad;
 }
